+++ src/common/utils/test/mux.go
@@ -1,0 +1,7 @@
+package test
+
+import "github.com/gorilla/mux"
+
+func NewMuxRouter() *mux.Router {
+	return mux.NewRouter()
+}
