+++ src/registryctl/handlers/mux.go
@@ -1,0 +1,7 @@
+package handlers
+
+import "github.com/gorilla/mux"
+
+func NewMuxRouter() *mux.Router {
+	return mux.NewRouter()
+}
